+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `/screenshot` endpoint: `format`, `quality`, `scale`, `selector`, `clip`, `color_scheme`, `wait_for`, `transparent` parameters ([#30])
+
 ## [0.4.1] - 2026-08-25
 
 ### Fixed